+++ tests/src/main/cpp/util/MockMslContext.cpp
@@ -151,8 +151,8 @@ int64_t MockMslContext::getTime()
 {
     struct timeval tp;
     gettimeofday(&tp, NULL);
-    uint64_t ms = tp.tv_sec * 1000ull + tp.tv_usec / 1000ull;
-    return ms;
+    uint64_t ms = static_cast<uint64_t>(tp.tv_sec) * 1000ull + static_cast<uint64_t>(tp.tv_usec) / 1000ull;
+    return static_cast<int64_t>(ms);
 }
 
 shared_ptr<ICryptoContext> MockMslContext::getMslCryptoContext()
